requirements.txt
beautifulsoup4==4.11.1
dash==2.6.1
dash_bootstrap_components==1.2.1
numpy==1.23.2
pandas==1.4.4
plotly==5.10.0
python_dateutil==2.8.2
requests==2.28.1
